Created handle winning and handle loosing actions
