+++ hero/script/HandleLoosingAction.cs
